+++ .github/workflows/deploy_prod.yml
@@ -66,11 +66,12 @@ jobs:
        - name: ECHO VAR
          run: |
            echo $MODULES_MODULE_SOURCE
-       - name: Check previous release
-         run: |
-           .github/checkPreviousChannelRelease.sh $MODULE_SOURCE_NAME ce $RELEASE_CHANNEL $MODULES_MODULE_TAG
        - uses: actions/checkout@v4
        - uses: deckhouse/modules-actions/setup@v1
+       - name: Check previous release
+         run: |
+           chmod +x .github/check_previous_channel_release.sh
+           .github/check_previous_channel_release.sh $MODULES_MODULE_NAME ce $RELEASE_CHANNEL $MODULES_MODULE_TAG
        - uses: deckhouse/modules-actions/deploy@v1
   
   job-EE:
@@ -85,11 +86,12 @@ jobs:
       - name: ECHO VAR
         run: |
            echo $MODULES_MODULE_SOURCE
-      - name: Check previous release
-        run: |
-          .github/checkPreviousChannelRelease.sh $MODULE_SOURCE_NAME ee $RELEASE_CHANNEL $MODULES_MODULE_TAG
       - uses: actions/checkout@v4
       - uses: deckhouse/modules-actions/setup@v1
+      - name: Check previous release
+        run: |
+          chmod +x .github/check_previous_channel_release.sh
+          .github/check_previous_channel_release.sh $MODULES_MODULE_NAME ee $RELEASE_CHANNEL $MODULES_MODULE_TAG
       - uses: deckhouse/modules-actions/deploy@v1
       
   job-FE:
@@ -104,9 +106,10 @@ jobs:
       - name: ECHO VAR
         run: |
            echo $MODULES_MODULE_SOURCE
-      - name: Check previous release
-        run: |
-          .github/checkPreviousChannelRelease.sh $MODULE_SOURCE_NAME fe $RELEASE_CHANNEL $MODULES_MODULE_TAG
       - uses: actions/checkout@v4
       - uses: deckhouse/modules-actions/setup@v1
+      - name: Check previous release
+        run: |
+          chmod +x .github/check_previous_channel_release.sh
+          .github/check_previous_channel_release.sh $MODULES_MODULE_NAME fe $RELEASE_CHANNEL $MODULES_MODULE_TAG
       - uses: deckhouse/modules-actions/deploy@v1
