+++ crds/doc-ru-localstorageclass.yaml
@@ -1,0 +1,56 @@
+spec:
+  versions:
+    - name: v1alpha1
+      schema:
+        openAPIV3Schema:
+          description: |
+            LocalStorageClass - это пользовательский ресурс Kubernetes, который определяет конфигурацию для Kubernetes Storage Class.
+          properties:
+            spec:
+              description: |
+                Описывает конфигурацию Kubernetes Storage Class.
+              properties:
+                reclaimPolicy:
+                  description: |
+                    Reclaim policy данного storage class'а. Может быть:
+                    - Delete (При удалении Persistent Volume Claim также удаляются Persistent Volume и связанное хранилище)
+                    - Retain (При удалении Persistent Volume Claim остаются Persistent Volume и связанное хранилище)
+                volumeBindingMode:
+                  description: |
+                    Binding mode для данного Storage class'а. Может быть:
+                    - Immediate (создает PV сразу же, как будет создан PVC)
+                    - WaitForFirstConsumer (создает PV только после того, как будет создан Pod для PVC)
+                lvm:
+                  description: |
+                    Поле описывает конфигурацию LVM.
+                  properties:
+                    type:
+                      description: |
+                        Тип девайса.
+                    lvmVolumeGroups:
+                      description: |
+                        LVMVolumeGroup ресурсы, на которых будут размещены Persistent Volume.
+                      items:
+                        properties:
+                          name:
+                            description: |
+                              Имя LVMVolumeGroup ресурса.
+                          thin:
+                            description: |
+                              Thin pool в выбранном LVMVolumeGroup ресурсе.
+                            properties:
+                              poolName:
+                                description: |
+                                  Имя выбранного Thin pool.
+            status:
+              description: |
+                Описывает текущую информацию о соответствующем Storage Class.
+              properties:
+                phase:
+                  description: |
+                    Текущее состояние Storage class. Может быть: 
+                    - Failed (в случае, если контроллер получил некорректную конфигурацию, или возникли ошибки в ходе выполнения операций)
+                    - Create (если все операции завершились успешно)
+                reason:
+                  description: |
+                    Дополнительная информация о состоянии Storage Class.
