+++ openapi/doc-ru-config-values.yaml
@@ -2,3 +2,13 @@ type: object
 properties:
   logLevel:
     description: Уровень логирования модуля.
+  dataNodes:
+    description: Настройки локальных томов csi на узлах с данными
+    properties:
+      nodeSelector:
+        description: |
+          То же, что и в параметре `spec.nodeSelector` подов в Kubernetes.
+                       
+          Если параметр опущен, локальный том csi будет размещен на всех узлах.
+
+          **Внимание!** Изменение этого параметра не приводит к перераспределению данных. Если узел с данными больше не соответствует «nodeSelector», данные на этом узле станут недоступными.
